requirements.txt
fastapi
uvicorn
httpx
watchfiles
pytest-asyncio